lx_boot_zone_ubuntu should check for non empty val
Currently `lx_boot_zone_ubuntu` manages /etc/resolv.conf if the user opts to manual manage the file, and not set resolvers and dns-domain in the zonecfg.

This is because `zone_attr()` will exit 0 even if the value is incorrect, `lx_boot` does not have this issue as it checks for a non empty value in stead. Applying the same change here makes ubuntu lx zones behave properly.

I tested by:
- boot a lx zone with resolvers and dns-domain set, copy /etc/resolv.conf
- replace lx_boot_zone_ubuntu
- boot a lx zone with resolvers and dns-domain set, diff /etc/resolv.conf /etc/resolv.conf.old --> no diff
- remove both resolvers and dns-domain from the zonecfg
- remove /etc/resolv.conf in the zone
- stop + boot the zone
- verify there is no /etc/resolv.conf create

The behavior before was a file with just the 'search ' line.